large code refactoring to break down the logic in smaller functions/methods
