+++ website/docs-scraper/config.mjs
@@ -1,0 +1,16 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
+import path from 'path';
+
+const TOKENS_JSON_FILEPATH =
+  '../packages/tokens/dist/docs/products/tokens.json';
+const FLIGHT_ICONS_JSON_FILEPATH = '../packages/flight-icons/catalog.json';
+
+export const config = {
+  distDocsFolder: path.resolve('dist/docs'), // path.resolve(__dirname, '../dist/docs');
+  tokensJsonFilePath: path.resolve(TOKENS_JSON_FILEPATH),
+  flightIconsJsonFilePath: path.resolve(FLIGHT_ICONS_JSON_FILEPATH),
+};
