+++ website/docs-scraper/parts/remarkRemoveCodeBlocks.mjs
@@ -1,0 +1,10 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
+import { remove } from 'unist-util-remove';
+
+export const remarkRemoveCodeBlocks = () => (tree) => {
+  remove(tree, (node) => node.type === 'code');
+};
