+++ website/docs-scraper/parts/stringifyChildNodes.mjs
@@ -1,0 +1,17 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
+export const stringifyChildNodes = (node) => {
+  const text = node.children.reduce((acc, child) => {
+    if ('children' in child) {
+      acc += stringifyChildNodes(child);
+    } else if ('value' in child) {
+      acc += ` ${child.value}`;
+    }
+    return acc;
+  }, '');
+
+  return text.trim();
+};
