+++ website/docs-scraper/parts/walkdir.mjs
@@ -1,3 +1,8 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
 import fs from 'fs-extra';
 import path from 'path';
